requirements.txt
praw
cohere
python-dotenv
tqdm
